Make OA optinal, improve Calibration, improve Serialisation
The Online Access Interface is now Optional (Default Off).
Therefore the RRL can now be used with recent Versions of Score-P like Score-P 5.0 or the upcoming Score-P 6.0.
Moreover, the Calibration is now more flexible and should work more Intel CPUs.
AMD is untested.
Finally, the Serialisation of the Q-Map has been improved, and is now more readable, as function names are given.